Correct Line Sequence from CRLF to LF. Move images to assets folder
